1. 添加生成IAP工具的模板工程 2. 移除hidtool 3. 添加将模板程序二进制数据转换成cpp字节数组的工具函数 4. 处理原有的警告，多数为数据类型大转小 5. 将crc函数的size参数从uint16_t改为uint32_t，考虑对bin整体进行crc时bin的大小超过65535 6. 工程配置修改，临时目录加一层项目名目录

+++ usb_fota/src/passthrough/passthrough.cpp
@@ -70,7 +70,7 @@ static void USB_SendData(uint16_t vid, uint16_t pid, int interfaceIndex, int ep,
 
 	// 向指定端点发送数据
 	int nActualBytes = 0;
-	nRet = libusb_bulk_transfer(pHandle, ep, (unsigned char*)data, size, &nActualBytes, 1000);
+	nRet = libusb_bulk_transfer(pHandle, ep, (unsigned char*)data, (int)size, &nActualBytes, 1000);
 	if (nRet < 0)
 	{
 		libusb_release_interface(pHandle, interfaceIndex);
@@ -82,7 +82,7 @@ static void USB_SendData(uint16_t vid, uint16_t pid, int interfaceIndex, int ep,
 	logger->AddLog("[info] %s\n", (std::stringstream() << "libusb_bulk_transfer(0x" << std::hex << (int)ep << std::dec << ") write size:[" << nActualBytes << "]").str().c_str());
 
 	unsigned char hexbuf[DATA_INPUT_BUFF_MAX_SIZE * 2] = { 0 };
-	utils::Hex2String(data, hexbuf, size);
+	utils::Hex2String(data, hexbuf, (int)size);
 	logger->AddLog("[info] %s\n", hexbuf);
 
 	// 释放第n个接口
@@ -130,7 +130,7 @@ static int USB_RecvData(uint16_t vid, uint16_t pid, int interfaceIndex, int ep, 
 
 	// 从指定端点接收数据
 	int nActualBytes = 0;
-	nRet = libusb_bulk_transfer(pHandle, ep, (unsigned char*)data, size, &nActualBytes, 10000);
+	nRet = libusb_bulk_transfer(pHandle, ep, (unsigned char*)data, (int)size, &nActualBytes, 10000);
 	if (nRet < 0)
 	{
 		libusb_release_interface(pHandle, interfaceIndex);
